rename tautology, adds :growth/robustness and :regression/noisier
[body]

tautology wasn't very descriptive and the word pollution (as in
namespace pollution) is more well known and descriptive.

Noisier was in response to me adding in a warning somewhere in a
codebase, and I realized that this is actually a stumbling block for
new users since they think somthing is wrong. On the other hand it
usually is accompanied by a growth item and the avoidance of a
regression to get there.

Robustness was added as a separate item from performance or fixation
because it does not fix something inherently erroneous in the
codebase (bug) and doesn't make the thing run with fewer
resources (performance), but rather handles more external failure
modes of the actual runtimes (i.e. leaky abstractions).

However, if the system you are building makes explicitly claims about
the robustness then this should be filed under (minimally) a fixation,
although a join fixation/robustness change is probably the most useful
there.

changes = [:growth/accretion, :growth/clarity, :breakage/rename]
domain = [:domains/spec]